Upgrade transpiler and expand TypeScript support
This update extends TypeScript support in the transpiler with new mapping features for various expressions and statements. Debugging is refactored from "console.log" in JavaScript to "debug" in LeekScript. The update also includes changes to the handling of initialization assignments in parameters and declarations. 'For-in' statements, conditional statements, and template literals are now supported. Minor adjustments have been made throughout for clarification and ease of use.